+++ Samples~/Sample3/Scripts/CategoryPopup.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+using Kuuasema.DataBinding;
+
+
+public class CategoryPopup : CategoryViewModel {
+
+ 
+}
+
+