+++ Samples~/Sample3/Scripts/DemoCanvas.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+using Kuuasema.DataBinding;
+
+public class DemoCanvas : DemoViewModel {
+    
+    protected override void Start() {
+        this.BindData(Demo.DataModel);
+    }
+}