+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+
+
+1.0.2
+==============
+- Added InputFieldTextViewModel
+- Added Sample1, Sample2, Sample3 
+- Improved assembly detection
+
 1.0.1
 ==============
 - Updated data binding to fetch models across all assemblies that reference it.